% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binband.R
\name{binband}
\alias{binband}
\title{Get linearly scaled variable based on optimal bins}
\usage{
binband(df, target = "value")
}
\arguments{
\item{df}{string, search term for the World Risk Poll questions}

\item{target}{Target column}
}
\value{
Banded data frame
}
\description{
Finds outliers and then bands between 0 and 1 on optimal bins of
non-outlier data
}
\examples{
binband(mtcars, target = "mpg")
}
