% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newIndex.R
\name{newIndex}
\alias{newIndex}
\title{Get linearly scaled variable based on optimal bins}
\description{
Finds outliers and then bands between 0 and 1 on optimal bins of
non-outlier data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-newIndex-new}{\code{newIndex$new()}}
\item \href{#method-newIndex-addIndicator}{\code{newIndex$addIndicator()}}
\item \href{#method-newIndex-impute}{\code{newIndex$impute()}}
\item \href{#method-newIndex-calculate_index}{\code{newIndex$calculate_index()}}
\item \href{#method-newIndex-clone}{\code{newIndex$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-new"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$new(name, start_year, end_year, peg_year)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-addIndicator"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-addIndicator}{}}}
\subsection{Method \code{addIndicator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$addIndicator(
  domain,
  variablename,
  source,
  ismorebetter,
  geocode,
  year,
  value,
  weight
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-impute"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-impute}{}}}
\subsection{Method \code{impute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$impute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-calculate_index"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-calculate_index}{}}}
\subsection{Method \code{calculate_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$calculate_index()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
