% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newIndex.R
\name{newIndex}
\alias{newIndex}
\title{Get linearly scaled variable based on optimal bins}
\description{
Finds outliers and then bands between 0 and 1 on optimal bins of
non-outlier data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-newIndex-new}{\code{newIndex$new()}}
\item \href{#method-newIndex-addIndicator}{\code{newIndex$addIndicator()}}
\item \href{#method-newIndex-viewStructure}{\code{newIndex$viewStructure()}}
\item \href{#method-newIndex-calculateIndex}{\code{newIndex$calculateIndex()}}
\item \href{#method-newIndex-viewERDiagram}{\code{newIndex$viewERDiagram()}}
\item \href{#method-newIndex-clone}{\code{newIndex$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-new"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$new(name, start_year, end_year)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-addIndicator"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-addIndicator}{}}}
\subsection{Method \code{addIndicator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$addIndicator(
  df,
  domain,
  ismorebetter,
  weight,
  manual_min_outlier_cutoff = NULL,
  manual_max_outlier_cutoff = NULL,
  banding_method = "optimal"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-viewStructure"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-viewStructure}{}}}
\subsection{Method \code{viewStructure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$viewStructure()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-calculateIndex"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-calculateIndex}{}}}
\subsection{Method \code{calculateIndex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$calculateIndex(export_to_db = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-viewERDiagram"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-viewERDiagram}{}}}
\subsection{Method \code{viewERDiagram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$viewERDiagram()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newIndex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-newIndex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newIndex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
