% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newindex.R
\name{newindex}
\alias{newindex}
\title{Get linearly scaled variable based on optimal bins}
\description{
Finds outliers and then bands between 0 and 1 on optimal bins of
non-outlier data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-newindex-new}{\code{newindex$new()}}
\item \href{#method-newindex-addIndicator}{\code{newindex$addIndicator()}}
\item \href{#method-newindex-structure}{\code{newindex$structure()}}
\item \href{#method-newindex-calcIndex}{\code{newindex$calcIndex()}}
\item \href{#method-newindex-viewERDiagram}{\code{newindex$viewERDiagram()}}
\item \href{#method-newindex-clone}{\code{newindex$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newindex-new"></a>}}
\if{latex}{\out{\hypertarget{method-newindex-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newindex$new(name, start_year, end_year)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newindex-addIndicator"></a>}}
\if{latex}{\out{\hypertarget{method-newindex-addIndicator}{}}}
\subsection{Method \code{addIndicator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newindex$addIndicator(
  df,
  domain,
  ismorebetter,
  weight,
  manual_min_outlier_cutoff = NULL,
  manual_max_outlier_cutoff = NULL,
  banding_method = "optimal"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newindex-structure"></a>}}
\if{latex}{\out{\hypertarget{method-newindex-structure}{}}}
\subsection{Method \code{structure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newindex$structure()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newindex-calcIndex"></a>}}
\if{latex}{\out{\hypertarget{method-newindex-calcIndex}{}}}
\subsection{Method \code{calcIndex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newindex$calcIndex()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newindex-viewERDiagram"></a>}}
\if{latex}{\out{\hypertarget{method-newindex-viewERDiagram}{}}}
\subsection{Method \code{viewERDiagram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newindex$viewERDiagram()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newindex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-newindex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newindex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
